+++ src/router/routes.js
@@ -15,6 +15,7 @@ import pgError404 from 'pages/Error404'
 import pgAdminCustomers from 'pages/admin/Customers'
 import pgTopup from 'pages/Topup'
 import pgMakePayment from 'pages/MakePayment'
+import pgAdminReceivePayment from 'pages/admin/ReceivePayment'
 
 const routes = [
   {
@@ -84,6 +85,10 @@ const routes = [
       {
         path: 'customers',
         component: pgAdminCustomers,
+      },
+      {
+        path: 'receivepayment',
+        component: pgAdminReceivePayment,
       },
     ],
   },